Add output to webpage, buttons, score counter
User selects from three buttons to play a round of RPS. Below the buttons,
the result of the previous games are displayed. Below that, a counter of
wins for both the player and computer is displayed.

When one party reaches five wins, the queue is cleared and the winner is
displayed in place of the round history and counter.

TODO: Make it look nicer, center the stuff and space it out properly.